try to run separately app and client
