+++ bin/open_browser.py
@@ -6,8 +6,6 @@ sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
 import config as conf
 
-
-
 url = f"http://localhost:{conf.SERVER_PORT}/"
 
 if sys.platform=='win32':
