+++ main.py
@@ -1,14 +1,16 @@
 import subprocess
 import argparse
 
+import config as conf
+
 def main():
     parser = argparse.ArgumentParser(
         description='Customize notification interval'
     )
-    parser.add_argument('--seconds', type=int, default=5, help='Notify every given seconds')
+    parser.add_argument('--seconds', type=int, default=conf.DEFAULT_INTERVAL, help='Notify every given seconds')
     args = parser.parse_args()
 
-    subprocess.run(['python', 'bin/orchestrator.py', str(args.seconds)])
+    subprocess.run(['python3', 'bin/orchestrator.py', str(args.seconds)])
 
 if __name__ == "__main__":
     main()
